moving logic out of route.ts
